+++ plone/tiles/tests/test_data.py
@@ -1,0 +1,40 @@
+# -*- coding: utf-8 -*-
+from plone.tiles.data import encode
+from plone.tiles.testing import PLONE_TILES_INTEGRATION_TESTING
+from zope import schema
+from zope.interface import Interface
+
+import unittest
+
+
+class IQuerySchema(Interface):
+
+    query = schema.List(
+        title=u"Search terms",
+        value_type=schema.Dict(value_type=schema.Field(),
+                               key_type=schema.TextLine()),
+        required=False
+    )
+
+    title = schema.TextLine(
+        title=u'Title'
+    )
+
+
+class TestEncode(unittest.TestCase):
+
+    layer = PLONE_TILES_INTEGRATION_TESTING
+
+    def test_encode_querystring_special(self):
+        data = {
+            'query': [{'i': 'Subject',
+                       'o': 'plone.app.querystring.operation.selection.any',
+                       'v': [u'äüö']}],
+            'title': u'Hello World'
+        }
+        self.assertEqual(
+            encode(data, schema=IQuerySchema),
+            ('query.i%3Arecords=Subject&query.o%3A'
+             'records=plone.app.querystring.operation.selection.any&'
+             'query.v%3Alist%3Arecords=%C3%A4%C3%BC%C3%B6&title=Hello+World')
+        )